refactor: move sponsored-send orchestration out of delegation domain
