+++ src/features/transfer/utils/sponsoredSend.ts
@@ -6,16 +6,15 @@ import { isAddress, isHex, type Address, type Hex } from 'viem';
 
 import { TransactionStatus, type NewTransaction } from '@/entities/transactions/transaction';
 import { getRemoteConfig } from '@/features/config/stores/remoteConfig';
+import { createDelegationPublicClient, SPONSORED_CALLS_REQUIREMENTS } from '@/features/delegation/calls';
+import { trackCallsExecution } from '@/features/delegation/callsExecutionTracking';
+import { resolveManagedExecutionFailure } from '@/features/delegation/managedExecutionFailure';
+import { predictSponsoredCallsExecution } from '@/features/delegation/sponsoredCalls';
+import { canUseDelegatedExecution, supportsDelegatedExecution } from '@/features/delegation/willDelegate';
 import { backendNetworksActions } from '@/features/network/stores/backendNetworksStore';
 import { type ChainId } from '@/features/network/types/backendNetworks';
 import { RainbowError } from '@/logger';
 import { execute, type Call, type ExecuteCallsResult, type ExecutionResult, type PreparedCallsExecution } from '@rainbow-me/delegation';
-
-import { createDelegationPublicClient, SPONSORED_CALLS_REQUIREMENTS } from './calls';
-import { trackCallsExecution } from './callsExecutionTracking';
-import { resolveManagedExecutionFailure } from './managedExecutionFailure';
-import { predictSponsoredCallsExecution } from './sponsoredCalls';
-import { canUseDelegatedExecution, supportsDelegatedExecution } from './willDelegate';
 
 // ============ Types ========================================================= //
